% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_readme_section.R
\name{add_readme_section}
\alias{add_readme_section}
\title{Add a Typewrite messages section in README
If the README.md in the present working directory does not contain a "# Typewrite Messages" heading, it is added in the end of it.}
\usage{
add_readme_section()
}
\description{
Add a Typewrite messages section in README
If the README.md in the present working directory does not contain a "# Typewrite Messages" heading, it is added in the end of it.
}
