% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_readme.R
\name{create_readme}
\alias{create_readme}
\title{Creates a README.md in the working directory if it doesn't exist}
\usage{
create_readme()
}
\description{
Creates a README.md in the working directory if it doesn't exist
}
