% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_to_readme_section.R
\name{log_to_readme_section}
\alias{log_to_readme_section}
\title{Log to a README by section}
\usage{
log_to_readme_section(section_name, entry)
}
\description{
This function differs from write_to_readme_section by the structure of the entry. This function writes the entries in a tab-separated format as seen in other logs while write_to_readme_section writes a single entry in a more human readable format.
}
