% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_TypewriteMessage.R
\name{log_TypewriteMessage}
\alias{log_TypewriteMessage}
\title{Log a Typewrite Message to log.md}
\usage{
log_TypewriteMessage(TypewriteMessage_obj, disable_rstudioapi = FALSE)
}
\description{
Log a Typewrite Message to log.md
}
