% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_PlainTypewriteMessage.R
\name{make_PlainTypewriteMessage}
\alias{make_PlainTypewriteMessage}
\title{Make a PlainTypewriteMessage}
\usage{
make_PlainTypewriteMessage(...)
}
\description{
This function fills the Plain slot of the TypewriteMessage object from the input
}
