% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_TypewriteOutput.R
\name{make_TypewriteOutput}
\alias{make_TypewriteOutput}
\title{Make the final output to be printed to console}
\usage{
make_TypewriteOutput(TypewriteLines_obj)
}
\description{
Make the final output to be printed to console
}
