% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_readme_section.R
\name{new_readme_section}
\alias{new_readme_section}
\title{Create a new section in README
If the README.md in the present working directory does not contain a section with the heading in the format of "# {section_name}", it is added in the end of it.}
\usage{
new_readme_section(section_name)
}
\arguments{
\item{section_name}{name of the section, excluding the "# " prefix}
}
\description{
Create a new section in README
If the README.md in the present working directory does not contain a section with the heading in the format of "# {section_name}", it is added in the end of it.
}
