% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_booklet_order.R
\name{pdf_to_booklet_order}
\alias{pdf_to_booklet_order}
\title{Converts a pdf file into a version with the pages rearranged for booklet-style printing}
\usage{
pdf_to_booklet_order(path_to_pdf)
}
\arguments{
\item{path_to_pdf}{path to a pdf file with a page number that is a multiple of 4}
}
\description{
Converts a pdf file into a version with the pages rearranged for booklet-style printing
}
