% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/press_enter.R
\name{press_enter}
\alias{press_enter}
\title{Stops the script and asks end-user to press [ENTER] to continue}
\usage{
press_enter()
}
\description{
Stops the script and asks end-user to press [ENTER] to continue
}
