% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch_error_as_na.R
\name{try_catch_error_as_na}
\alias{try_catch_error_as_na}
\title{Skip error messages, records NA, and continues to loop on the expression}
\usage{
try_catch_error_as_na(expr)
}
\arguments{
\item{expr}{expression}
}
\description{
Skip error messages, records NA, and continues to loop on the expression
}
