% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_catch_error_as_null.R
\name{try_catch_error_as_null}
\alias{try_catch_error_as_null}
\title{Skip error messages and continues to loop on the expression}
\usage{
try_catch_error_as_null(expr)
}
\arguments{
\item{expr}{expression}
}
\description{
Skip error messages and continues to loop on the expression
}
