% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typewrite.R
\name{typewrite}
\alias{typewrite}
\title{Returns a phrase with formatting}
\usage{
typewrite(..., tabs = 0, c_return_lines = 1)
}
\arguments{
\item{...}{words in the phrase}

\item{tabs}{number of tab spacing desired. defaults to 1.}

\item{c_return}{number of lines wanted after the phrase. defaults to 1.}
}
\description{
Returns a phrase with formatting
}
