% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typewrite_bold.R
\name{typewrite_bold}
\alias{typewrite_bold}
\title{Returns a phrase with formatting}
\usage{
typewrite_bold(..., tabs = 0, line_number = 0, log = FALSE)
}
\arguments{
\item{...}{words in the phrase}

\item{tabs}{number of tab spacing desired. defaults to 1.}

\item{line_number}{number of lines wanted after the phrase. defaults to 1.}
}
\description{
Returns a phrase with formatting
}
