% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typewrite_error.R
\name{typewrite_error}
\alias{typewrite_error}
\title{Returns a ERROR message in console}
\usage{
typewrite_error(..., tabs = 0, line_number = 0, add_to_readme = FALSE)
}
\arguments{
\item{...}{words in the phrase}

\item{tabs}{number of tab spacing desired. defaults to 1.}

\item{c_return}{number of lines wanted after the phrase. defaults to 1.}
}
\description{
This function is like the typewrite() function, except there is an "ERROR:" prefix attached to the console in red.
}
