% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typewrite_note.R
\name{typewrite_note}
\alias{typewrite_note}
\title{Returns a NOTE message in console}
\usage{
typewrite_note(..., tabs = 0, line_number = 1, add_to_readme = TRUE)
}
\arguments{
\item{...}{words in the phrase}

\item{tabs}{number of tab spacing desired. defaults to 1.}

\item{c_return}{number of lines wanted after the phrase. defaults to 1.}
}
\description{
This function is like the typewrite() function, except there is an "ERROR:" prefix attached to the console in red.
}
