% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typewrite_progress.R
\name{typewrite_progress}
\alias{typewrite_progress}
\title{Typewrite Progress of a Loop}
\usage{
typewrite_progress(iteration, total, file = "", append = TRUE)
}
\arguments{
\item{iteration_n}{integer of iteration of the loop}

\item{.input}{dataframe, tibble, list, or vector that is being iterated over.}

\item{timestamped}{If TRUE, a timestamp of the Sys.time is included in the line surrounded in square brackets}
}
\description{
Typewrite Progress of a Loop
}
