% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Classes.R, R/set_TypewriteLines_class.R
\docType{class}
\name{TypewriteLines-class}
\alias{TypewriteLines-class}
\alias{TypewriteLines}
\title{Set the TypewriteLines Class}
\description{
Set the TypewriteLines Class

Set the TypewriteLines Class
}
