% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Classes.R, R/set_TypewriteMessage_class.R
\docType{class}
\name{TypewriteMessage-class}
\alias{TypewriteMessage-class}
\alias{TypewriteMessage}
\title{Set TypewriteMessage Class}
\description{
Set TypewriteMessage Class

Set TypewriteMessage Class
}
