% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Classes.R,
%   R/set_TypewriteMessageLog_class.R
\docType{class}
\name{TypewriteMessageLog-class}
\alias{TypewriteMessageLog-class}
\alias{TypewriteMessageLog}
\title{Set TypewriteMessageLog Class}
\description{
Set TypewriteMessageLog Class

Set TypewriteMessageLog Class
}
