% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_readme_section.R
\name{write_to_readme_section}
\alias{write_to_readme_section}
\title{Writes to a README by section}
\usage{
write_to_readme_section(section_name, entry)
}
\description{
A vector of length 1 or greater is added with a timestamp under a the header determined by the provided section_name to the README.md of the working directory.
}
