% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_typewrite_to_readme.R
\name{write_typewrite_to_readme}
\alias{write_typewrite_to_readme}
\title{Writes a typewritten message to README}
\usage{
write_typewrite_to_readme(typewrite_message)
}
\description{
The typewrite message is added with a timestamp under a #Typewrite Messages section in the README.md of the working directory.
}
