% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typewrite.R
\name{typewrite}
\alias{typewrite}
\title{Typewrite}
\usage{
typewrite(
  ...,
  tabs = 0,
  line_number = 0,
  timepunched = TRUE,
  file = "",
  append = TRUE,
  log = FALSE
)
}
\arguments{
\item{...}{Strings to concatenate into a single message}

\item{tabs}{Number of tabs to indent the message before it is printed to the console, Default: 0}

\item{line_number}{Number of lines to skip before and after the message is printed to the console, Default: 0}

\item{timepunched}{Should the message be prefixed with a bracketed timestamp?, Default: TRUE}

\item{file}{Argument passed to the cat function if the message should also be saved to a file, Default: ''}

\item{append}{Argument passed to the cat function if the message should be appended to the file argument, Default: TRUE}
}
\description{
Typewrite
}
\seealso{
\code{\link[crayon]{strip_style}}
}
